fix: fix bot command message and hook failure
